fix: 修复财富排行榜货币字段潜在的数字溢出问题
- 将财富排行榜中的灵石与银两 SQL 类型转换从 int 修改为 bigint，防止高资产账号数值溢出。
- 在回归测试中增加强制类型断言，确保相关货币字段不再被错误强转为 int4，预防后续代码回归。

+++ server/src/services/rankService.ts
@@ -188,8 +188,8 @@ const loadWealthRanks = async (limit: number): Promise<WealthRankRow[]> => {
         title,
         avatar,
         realm,
-        COALESCE(spirit_stones, 0)::int AS "spiritStones",
-        COALESCE(silver, 0)::int AS silver
+        COALESCE(spirit_stones, 0)::bigint AS "spiritStones",
+        COALESCE(silver, 0)::bigint AS silver
       FROM characters
       WHERE nickname IS NOT NULL AND nickname <> ''
       ORDER BY rank
